Wellynton_Spagnol/infratech/src/main/java/org/api/celcoin/infra/celcoin/dto/CelcoinPixResponseDTO.java
package org.api.celcoin.infra.celcoin.dto;


import jakarta.json.bind.annotation.JsonbProperty;

public class CelcoinPixResponseDTO {
    private Long transactionId;
    @JsonbProperty(value = "emvqrcps")
    private String qrCode;

    public Long getTransactionId() {
        return transactionId;
    }

    public void setTransactionId(Long transactionId) {
        this.transactionId = transactionId;
    }

    public String getQrCode() {
        return qrCode;
    }

    public void setQrCode(String qrCode) {
        this.qrCode = qrCode;
    }
}
